fix: Fixed the errors deriving from earlier changes, it gives me 44 error from -Wswitch and I have no idea why or how I should manage all of the inputs from the user

+++ include/SimParams.hpp
@@ -5,7 +5,7 @@
 namespace flock {
 
 struct SimParams {
-  int BoidNum{0};
+  long unsigned int BoidNum{0};
   double s{0.};
   double a{0.};
   double c{0.};
